Implement SmallArray
SmallArray is an SBO-optimized Array which addresses the same use case
as KJ_STACK_ARRAY currently does.

KJ_STACK_ARRAY has the advantage that it uses a VLA on gcc, which may
allow for more optimal use of stack space for some use cases. The
implementation of KJ_STACK_ARRAY for Clang, however, does not use VLAs,
meaning we have not benefited from this feature for a long time (our
downstream codebases all use Clang). This limitation was reportedly
originally due to Clang's lack of support for non-trivial types in VLAs,
although nowadays it appears to support them. At any rate, VLAs are
obviously incompatible with coroutines, and downstream code uses
KJ_STACK_ARRAY in coroutines.

KJ_STACK_ARRAY has the disadvantage that it is a macro whose arguments
are fairly easy to mix up. I believe SmallArray's API design is safer
and more obvious to get right, and would like to begin using it in
preference to KJ_STACK_ARRAY.