Fix transfer of headers when suspending a request.
We really need to transfer the original headers object, not clone it. Usually the headers all point into the original header buffer, but not necessarily always. The headers could have one or more headers backed by its `ownedStrings`.

Luckily as we are dismantling the connection anyway, we're free to take the headers with us.