Fix HashIndex::reserve to actually reserve enough space
This brings it closer in line to the logic in insert() that decides when
to grow the index, making sure to leave enough room so that we won't
have to grow/rehash again if `size` entries are inserted. It looks like
calling rehash(size * 2) was probably always the intent and it was just
a mistake that we weren't doing that? I'd be interested to know why not
if just calling rehash(size) was really what we wanted.

I've also accounted for erasedCount here as well, again to bring things
more in line with the resizing logic in insert().