Remove StringPtr::attach()
This is a footgun (very easy to construct a ConstString and use it after the
underlying String gets deallocated, resulting in UaF), and usage appears to be
very limited (used just two times in Cloudflare downstream repos). The version
taking arguments is harder to use incorrectly, but appears completely unused,
remove it too.