Add `tryDowncast()`, a nice wrapper for dynamic_cast.
This is actually identical to the existing `dynamicDowncastIfAvailable()`, except that it is ONLY Defined if RTTI is enabled, whereas `dynamicDowncastIfAvailable()` returns `none` when RTTI is disabled. `dynamicDowncastIfAvailable()`'s behavior makes sense when using RTTI to detect optimization opportunities, but the Wokers Runtime now has a few places that use dynamic casting that isn't just for optimization, so `dynamicDowncastIfAvailable()` feels wrong to use there.