capnp-mode.el: fix comment configuration
The existing approach only works if you use `comment-dwim`, but that
function is just a wrapper around lower-level ones that actually consume
the various `comment-` vars. So, if you don't use `comment-dwim`, you
are not helped by capnp-mode in this regard.

So, drop the existing approach and just set an appropriate value for
`comment-start` in all capnp-mode buffers. This is typically how major
modes configure this kind of thing. I can't say why it was done a
different way to start with. (The configuration of `comment-end` in
`capnp-comment-dwim` was redundant, as `""` is the default value of this
var.)